создает модуляцию js
