+++ js/data.js
@@ -1,0 +1,17 @@
+const TITLES = ['Большое монументальное парадное здание, выделяющееся своей архитектурой, являлось первоначально резиденцией царствующих владетельных лиц и высшей знати', 'Превосходная квартира с роскошным дизайном интерьера расположена в центре Токио', 'Роскошные апартаменты с видом на город, в 2 км от храма Омори Хачиман и в 2,2 км от святилища Мива-Ицукусим', 'Прекрасная квартира оформленная в стиле минимализм', 'Дом для отпуска с 3 спальнями и патио находится в Токио, всего в 1,1 км от краеведческого музея Сугинами и парка Вадабери'];
+
+const TYPES = ['palace', 'flat', 'house ', 'bungalow'];
+
+const CHECKIN = ['12:00', '13:00', '14:00'];
+
+const CHECKOUT = ['12:00', '13:00', '14:00'];
+
+const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+
+const DESCRIPTIONS = ['К услугам гостей сауна и душ с функцией ароматерапии. Желающие смогут заказать массажные и косметические процедуры. Гостям предоставляются услуги персонального тренера. За дополнительную плату организован трансфер от/до аэропорта на лимузине Rolls-Royce. Стойка регистрации отеля открыта круглосуточно. На территории обустроена платная парковка.', 'Среди популярных достопримечательностей рядом находятся — парк Хикава, Гейно Каденша и храм Кумано.', 'Популярные достопримечательности вблизи включают краеведческий музей Сетагая, дополнительное здание Мукаи Хункичи и храм Джокко-дзи. Расстояние до международного аэропорта Токио Ханэда составляет 32 км.', 'Популярные достопримечательности вблизи включают храм Инари Кио, музей Кореи и парк Окубо. Расстояние до токийского международного аэропорта Ханэда составляет 17 км. За дополнительную плату организуется трансфер от/до аэропорта.'];
+
+const PHOTOS = ['http://o0.github.io/assets/images/tokyo/hotel1.jpg', 'http://o0.github.io/assets/images/tokyo/hotel2.jpg', 'http://o0.github.io/assets/images/tokyo/hotel3.jpg'];
+
+const SIMILAR_HOUSE_INFO_COUNT = 10;
+
+export{TITLES, TYPES, CHECKIN, CHECKOUT, FEATURES, DESCRIPTIONS, PHOTOS, SIMILAR_HOUSE_INFO_COUNT};
